Removed garbled text accidentally added in v0.91 from Civilopedia entries for buildings.
Replaced the Great Spies named after BtS playtesters with Great Spies from etiennefd's list here: https://github.com/dguenms/Dawn-of-Civilization/pull/199
[advc.005b]

PerfectMongoose:
* Added some code that encourages at least one starting plot per continent to be placed inland. The original script almost never produces inland starts. [advc.021b]
* Use AI_foundValue for preselecting starting plots (instead of the much simpler custom "PotentialValue" function). [advc.021b]

+++ CvGameCoreDLL/WarAndPeaceCache.cpp
@@ -647,7 +647,7 @@ void WarAndPeaceCache::sortCitiesByAttackPriority() {
 	PROFILE_FUNC();
 	/*  Selection sort b/c I want to factor in city areas. An invading army tends
 		to stay in one area until all cities there are conquered. */
-	for(size_t i = 0; i < v.size() - 1; i++) {
+	for(int i = 0; i < ((int)v.size()) - 1; i++) {
 		CvCity* cvc = NULL;
 		if(i > 0)
 			cvc = v[i - 1]->city();
